Added pre-submitted sub statuses to test data generator
