+++ app/services/testdata/StatusGeneratorFactory.scala
@@ -25,6 +25,10 @@ object StatusGeneratorFactory {
     status match {
       case "REGISTERED" => RegisteredStatusGenerator
       case ApplicationStatuses.Created => CreatedStatusGenerator
+      case "IN_PROGRESS_PERSONAL_DETAILS" => InProgressPersonalDetailsStatusGenerator
+      case "IN_PROGRESS_SCHEME_PREFERENCES" => InProgressSchemePreferencesStatusGenerator
+      case "IN_PROGRESS_ASSISTANCE_DETAILS" => InProgressAssistanceDetailsStatusGenerator
+      case "IN_PROGRESS_QUESTIONNAIRE" => InProgressAssistanceDetailsStatusGenerator
       case ApplicationStatuses.Submitted => SubmittedStatusGenerator
       case ApplicationStatuses.OnlineTestInvited => OnlineTestInvitedStatusGenerator
       case ApplicationStatuses.OnlineTestStarted => OnlineTestStartedStatusGenerator
